[val-7] Fixed model types and exceptions

+++ src/main/java/org/ticketshop/model/Location.java
@@ -21,7 +21,7 @@ public class Location {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private BigDecimal id;
+    private Long id;
 
     @NotNull
     @Column(name = "longitude", nullable = false)
